Add the Picard features to ReleaseInfoFragment.kt
